+++ packages/docusaurus-theme-redoc/src/theme/ApiSchema/index.ts
@@ -1,0 +1,3 @@
+import ApiSchema from './ApiSchema';
+
+export default ApiSchema;