+++ website/src/pages/examples/index.jsx
@@ -8,7 +8,7 @@ import styles from './styles.module.css';
 const sections = [
   {
     title: <>Using ApiDoc Component</>,
-    link: '/custom-page/',
+    link: '/examples/custom-page/',
     description: (
       <>
         A page using ApiDoc component in{' '}
@@ -20,7 +20,7 @@ const sections = [
   },
   {
     title: <>Using Redoc Component</>,
-    link: '/custom-layout/',
+    link: '/examples/custom-layout/',
     description: (
       <>
         A page with custom Layout using Redoc component in{' '}
@@ -32,7 +32,7 @@ const sections = [
   },
   {
     title: <>Using Spec URL</>,
-    link: '/using-spec-url/',
+    link: '/examples/using-spec-url/',
     description: (
       <>
         A page made automatically with{' '}
@@ -44,7 +44,7 @@ const sections = [
   },
   {
     title: <>Using YAML Spec</>,
-    link: '/using-spec-yaml/',
+    link: '/examples/using-spec-yaml/',
     description: (
       <>
         A page made automatically with{' '}