+++ packages/docusaurus-theme-redoc/CHANGELOG.md
@@ -1,5 +1,17 @@
 # docusaurus-theme-redoc
 
+## 0.5.0-next.0
+
+### Minor Changes
+
+- [`0fbefef`](https://github.com/rohit-gohri/redocusaurus/commit/0fbefef62247f1153102254613a120793a329861) Thanks [@rohit-gohri](https://github.com/rohit-gohri)! - Allow importing single model defenition inside mdx files - #46
+
+* [`12d7a26`](https://github.com/rohit-gohri/redocusaurus/commit/12d7a261c2736860fd48d77d5727b0f7127b807b) Thanks [@rohit-gohri](https://github.com/rohit-gohri)! - Add support for SSR - #42
+
+### Patch Changes
+
+- [`e0856fa`](https://github.com/rohit-gohri/redocusaurus/commit/e0856faccaae7ef65768ef7c3453d8a272800562) [#86](https://github.com/rohit-gohri/redocusaurus/pull/86) Thanks [@rohit-gohri](https://github.com/rohit-gohri)! - Fix common types not being exported
+
 ## 0.4.6
 
 ### Patch Changes